# src/main/java/com/ilbeol/staff/domain/model/auth/AuthResponse.java
package com.ilbeol.staff.domain.model.auth;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
public class AuthResponse {
    private String token;
}