# src/main/java/com/ilbeol/staff/infraestructure/controller/AuthController.java
package com.ilbeol.staff.infraestructure.controller;

import com.ilbeol.staff.domain.model.auth.AuthRequest;
import com.ilbeol.staff.domain.model.auth.AuthResponse;
import com.ilbeol.staff.domain.model.auth.RegisterRequest;
import com.ilbeol.staff.domain.port.input.auth.AuthService;
import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/auth")
@RequiredArgsConstructor
public class AuthController {

    private final AuthService authService;

    @PostMapping("/register")
    public ResponseEntity<AuthResponse> register(@RequestBody RegisterRequest request) {
        return ResponseEntity.ok(authService.register(request));
    }

    @PostMapping("/login")
    public ResponseEntity<AuthResponse> login(@RequestBody AuthRequest request) {
        return ResponseEntity.ok(authService.login(request));
    }
}